feat: Make map full screen and display timeline in the bottom sheet (#6)
* chore: Skeleton changes for making messages view draggable

* chore: Operate on a single data class to avoid wrong state handling

* chore: Move to standard BottomSheetScaffold from material lib.

* chore: Display messages view initially after entering the room

* chore: Resize bottom sheet respectively to the keyboard

* chore: Remove text compose overlay

* chore: Code clean-up

* chore: Add animation label

* chore: Further code clean-up

* chore: Revert redundant change

* chore: Format code

* chore: Format code

* chore: Disable swipe when lazyColumn is scrolled
