+++ features/messages/impl/src/main/kotlin/io/element/android/features/messages/impl/MessagesState.kt
@@ -51,5 +51,5 @@ data class MessagesState(
     val appName: String,
     val pinnedMessagesBannerState: PinnedMessagesBannerState,
     val eventSink: (MessagesEvents) -> Unit,
-    val isMessagesCollapsed: Boolean = false
+    val showMapView: Boolean
 )
