+++ features/messages/impl/src/main/kotlin/io/element/android/features/messages/impl/crypto/identity/MessagesViewWithIdentityChangePreview.kt
@@ -41,7 +41,6 @@ internal fun MessagesViewWithIdentityChangePreview(
         onJoinCallClick = {},
         onViewAllPinnedMessagesClick = {},
         knockRequestsBannerView = {},
-        onShowMapClick = TODO(),
         modifier = TODO(),
         forceJumpToBottomVisibility = TODO(),
         mapRealtimeState = TODO()
